pyproject.toml
[project]
name = "agents-for-david"
version = "2.0.0"
description = "Термодинамическая система с детерминированной фильтрацией"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
	"pandas>=2.2.0",
	"ipykernel>=6.29.0",
	"python-dotenv>=1.0.0",
	"pydantic-ai==1.0.8",
	"anthropic>=0.40.0",
	"tabulate>=0.9.0",
	"psutil>=5.9.0",
	"aiohttp>=3.9.0",
	"scipy>=1.16.2",
	"python-telegram-bot>=20.7",
	"asyncio-throttle>=1.0.2",
]

[project.scripts]
agents-for-david = "main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
